Added checks for login in DomainContact

+++ src/Domain/DomainContact.java
@@ -17,12 +17,12 @@ public class DomainContact implements IDomainContact {
     }
 
     private DomainContact() {
-        this.currentUser = null; //new SocialWorker(1000950000, "Morten", 'M', "10-01-0000", "Hejsavej", 88888888, "hej@nal.mail", 6, "Loc", "1234567");
+        this.currentUser = null;
     }
 
     @Override
     public void createCaseRequest(long citizenCPR, String desc, boolean isMessageClear, boolean isCarePackage, boolean isRehousingPackage, String requestPerson, boolean isCitizenInformed, String citizenName, char citizenGender, String citizenBirthdate, String citizenAddress, Integer citizenPhoneNr, String citizenMail) {
-        if (currentUser instanceof CaseEmployee) {
+        if (userLoggedIn() && currentUser instanceof CaseEmployee) {
             CaseEmployee caseEmployee = (CaseEmployee) currentUser;
             caseEmployee.createCaseRequest(PersistanceContact.getInstance().getNewCaseRequestID(), currentUser.getId(), citizenCPR, desc, isMessageClear, isCarePackage, isRehousingPackage, requestPerson, isCitizenInformed, citizenName, citizenGender, citizenBirthdate, citizenAddress, citizenPhoneNr, citizenMail);
         }
@@ -32,7 +32,7 @@ public class DomainContact implements IDomainContact {
 
     @Override
     public void createCase(int caseRequestID, String nextAppointment, String guardianship, String personalHelper, String personalHelperPowerOfAttorney, String citizenRights, boolean citizenInformedElectronic, boolean consent, String consentType, String[] collectCitizenInfo, String specialCircumstances, String differentCommune) {
-        if (currentUser instanceof SocialWorker) {
+        if (userLoggedIn() && currentUser instanceof SocialWorker) {
             SocialWorker socialWorker = (SocialWorker) currentUser;
             socialWorker.saveCase(PersistanceContact.getInstance().getNewCaseID(), caseRequestID, nextAppointment, guardianship, personalHelper, personalHelperPowerOfAttorney, citizenRights, citizenInformedElectronic, consent, consentType, collectCitizenInfo, specialCircumstances, differentCommune);
         }
@@ -42,7 +42,7 @@ public class DomainContact implements IDomainContact {
     
     @Override
     public void saveEditedCase(int caseID, int caseRequestID, String nextAppointment, String guardianship, String personalHelper, String personalHelperPowerOfAttorney, String citizenRights, boolean citizenInformedElectronic, boolean consent, String consentType, String[] collectCitizenInfo, String specialCircumstances, String differentCommune) {
-        if (currentUser instanceof SocialWorker) {
+        if (userLoggedIn() && currentUser instanceof SocialWorker) {
             SocialWorker socialWorker = (SocialWorker) currentUser;
             socialWorker.saveCase(caseID, caseRequestID, nextAppointment, guardianship, personalHelper, personalHelperPowerOfAttorney, citizenRights, citizenInformedElectronic, consent, consentType, collectCitizenInfo, specialCircumstances, differentCommune);
         }
@@ -53,7 +53,7 @@ public class DomainContact implements IDomainContact {
     @Override
     public void addEmployee(long CPR, String name, char gender, String birthdate, String Address,
             Integer phoneNr, String mail, String username, String password, int positionNumber) {
-        if (currentUser instanceof Admin) {
+        if (userLoggedIn() && currentUser instanceof Admin) {
             Admin admin = (Admin) currentUser;
             admin.addEmployee(CPR, name, gender, birthdate, Address, phoneNr, mail, PersistanceContact.getInstance().getNewEmployeeID(), username, password, positionNumber);
         }
@@ -63,7 +63,7 @@ public class DomainContact implements IDomainContact {
 
     @Override
     public void deleteEmployee(int employeeID) {
-        if (currentUser instanceof Admin) {
+        if (userLoggedIn() && currentUser instanceof Admin) {
             Admin admin = (Admin) currentUser;
             admin.deleteEmployee(employeeID);
         }
@@ -91,6 +91,11 @@ public class DomainContact implements IDomainContact {
 
     @Override
     public IPerson getPerson(long CPR) {
+        if (!userLoggedIn())
+        {
+            printUnauthorizedAccess("getPerson");
+            return null;
+        }
         PersistanceContact PS = PersistanceContact.getInstance();
         PS.logAction(currentUser.getId(), LogAction.GET_EMPLOYEE, "User requested for a person with CPR: " + CPR);
         return PS.getPerson(CPR);
@@ -99,6 +104,8 @@ public class DomainContact implements IDomainContact {
     public Employee getCurrentUser() {
         return currentUser;
     }
+    
+    private boolean userLoggedIn() { return currentUser != null;}
     
     private void printUnauthorizedAccess(String methodName) {
         System.out.println("User not allowed to perform command: " + methodName);
